finished chapter about Gates

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -63,8 +64,11 @@ class AttendeeController extends Controller
   /**
    * Remove the specified resource from storage.
    */
-  public function destroy(string $event, Attendee $attendee)
+  public function destroy(Event $event, Attendee $attendee)
   {
+    if (Gate::denies('delete-attendee', [$event, $attendee])) {
+      abort(403, 'You are not authorized to update this Event');
+    }
     $attendee->delete();
 
     return response(status: 204);
